backend/requirements.txt
alembic~=1.16.5
python-dotenv~=1.1.1
psycopg2
fastapi[standard]~=0.116.1
SQLAlchemy~=2.0.43